+++ backend/src/main/java/com/bugenzhao/bookstore_backend/entity/AuthedUser.java
@@ -5,14 +5,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class AuthedUser {
     @JsonProperty("user_id")
-    public int userId;
-
+    public long userId;
     public String username;
-
     @JsonProperty("user_type")
     public int userType;
 
-    public AuthedUser(int userId, String username, int userType) {
+    public AuthedUser(long userId, String username, int userType) {
         this.userId = userId;
         this.username = username;
         this.userType = userType;
