+++ backend/src/main/java/com/bugenzhao/bookstore_backend/entity/db/UserAuth.java
@@ -4,18 +4,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class UserAuth {
     @JsonProperty("user_id")
-    public int userId;
-
+    public long userId;
     public String username;
-
     public String password;
-
     @JsonProperty("user_type")
     public int userType;
-
     public boolean banned;
 
-    public UserAuth(int userId, String username, String password, int userType, boolean banned) {
+    public UserAuth(long userId, String username, String password, int userType, boolean banned) {
         this.userId = userId;
         this.username = username;
         this.password = password;
