+++ backend/src/main/java/com/bugenzhao/bookstore_backend/entity/db/rowmapper/UserAuthRowMapper.java
@@ -10,7 +10,7 @@ import com.bugenzhao.bookstore_backend.entity.db.UserAuth;
 public class UserAuthRowMapper implements RowMapper<UserAuth> {
     @Override
     public UserAuth mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new UserAuth(rs.getInt("user_id"), rs.getString("username"), rs.getString("password"),
+        return new UserAuth(rs.getLong("user_id"), rs.getString("username"), rs.getString("password"),
                 rs.getInt("user_type"), rs.getBoolean("banned"));
     }
 }
