+++ backend/src/main/java/com/bugenzhao/bookstore_backend/service/impl/BookServiceImpl.java
@@ -19,7 +19,7 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Optional<Book> findById(int bookId) {
+    public Optional<Book> findById(long bookId) {
         var result = jdbcTemplate.query("select * from books where id = ?", new BookRowMapper(), bookId);
         return result.stream().findFirst();
     }
